Refactor login redirect to use configurable OAuth URL
This commit updates the login redirect to use the `discordOAuthUrl` field in the `Config` object, rather than hard-coding the URL. This allows the OAuth URL to be easily changed without having to modify the code. The `discordOAuthUrl` field is added to the `Config` object, and the `load` method in the `Pages` class is updated to accept a `Config` object as an argument.