Remove unused @SuppressWarnings annotation in OAuth2Handler class
This commit removes the unused @SuppressWarnings annotation in the OAuth2Handler class. This annotation was originally added to suppress warnings for a method that was declared but not yet used. However, the method is now being used in the Pages class, so the annotation is no longer necessary and can be safely removed. This will help to keep the code clean and free of unnecessary annotations.