Update addColumn method calls to include table variable
This change refactors the code in the checkColumns method to use a table
variable in the calls to the SQLHandler.addColumn method. This makes the code
easier to read and also allows for future-proofing by allowing for the
possibility of adding additional tables in the future without having to
modify the code too much.