Refactor profile page data processing
This commit refactors the way in which the data for the profile page is processed. The data is now read from the session attributes in the Javalin `Context` object and inserted into the HTML using the `processPage` method in the `Pages` class. Labels have also been added to the profile page fields for clarity.